remove proxy component
